requirements.txt
asgiref==3.7.2
black==24.1.0
Django==3.2.23
pandas==2.1.4
psycopg2-binary==2.9.9
psycopg2-pool==1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
python-dotenv==1.0.0
mmh3==3.1.0
numpy==1.26.1
